+++ fabric/src/main/java/com/gitlab/srcmc/rctapi/fabric/FabricCommon.java
@@ -1,6 +1,6 @@
 /*
  * This file is part of Radical Cobblemon Trainers API.
- * Copyright (c) 2024, HDainester, All rights reserved.
+ * Copyright (c) 2025, HDainester, All rights reserved.
  *
  * Radical Cobblemon Trainers API is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -18,13 +18,27 @@
 package com.gitlab.srcmc.rctapi.fabric;
 
 import com.gitlab.srcmc.rctapi.ModCommon;
+import com.gitlab.srcmc.rctapi.ModRegistries;
+import com.gitlab.srcmc.rctapi.commands.arguments.BattleEndCommandMapArgument;
+import com.gitlab.srcmc.rctapi.commands.arguments.BattleRulesArgument;
 
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.command.v2.ArgumentTypeRegistry;
 
 public final class FabricCommon implements ModInitializer {
     @Override
     public void onInitialize() {
         ModCommon.init();
         // com.gitlab.srcmc.rctapi.example.ExampleMod.init(); // uncomment for example
+
+        ArgumentTypeRegistry.registerArgumentType(
+            ModRegistries.ArgumentTypes.BATTLE_RULES.getId(),
+            BattleRulesArgument.class,
+            ModRegistries.ArgumentTypes.BATTLE_RULES.get());
+
+        ArgumentTypeRegistry.registerArgumentType(
+            ModRegistries.ArgumentTypes.BATTLE_END_COMMAND_MAP.getId(),
+            BattleEndCommandMapArgument.class,
+            ModRegistries.ArgumentTypes.BATTLE_END_COMMAND_MAP.get());
     }
 }